Method names updated for consistency, method functionality added

+++ src/main/java/com/danekblom/lexicon/finalprojectrest/services/AudioServiceImpl.java
@@ -2,17 +2,22 @@ package com.danekblom.lexicon.finalprojectrest.services;
 
 import com.danekblom.lexicon.finalprojectrest.data.AudioRepository;
 import com.danekblom.lexicon.finalprojectrest.model.Audio;
+import com.danekblom.lexicon.finalprojectrest.model.DocumentationResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @Service
 public class AudioServiceImpl {
 
     @Autowired
     private AudioRepository audioRepository;
+
+    private Map<String, String> error = new HashMap<>();
 
     public AudioServiceImpl() {
     }
@@ -21,21 +26,39 @@ public class AudioServiceImpl {
         return audioRepository.save(newAudio);
     }
 
-    public void updateAudio(Audio audioItemToUpdate) {
-        audioRepository.save(audioItemToUpdate);
+    public Audio updateAudioItem(Audio audioItemToUpdate) {
+        return audioRepository.save(audioItemToUpdate);
     }
 
-    public void deleteAudio(Audio audioItemToDelete) {
-        audioRepository.delete(audioItemToDelete);
+    public void deleteAudioItem(Audio audioItemToDelete) {
+        if (audioRepository.findAll().contains(audioItemToDelete)) {
+            audioRepository.delete(audioItemToDelete);
+        }
+        else {
+            new DocumentationResponse("Item Not Found.");
+        }
+    }
+
+    public List<Audio> listAllAudioItems() {
+
+        return audioRepository.findAll();
     }
 
     public Audio findById(Integer id) {
+
         for (Audio foundAudioItem : audioRepository.findAll())
         {
             if (foundAudioItem.getId() == id) {
                 return foundAudioItem;
             }
         }
+
+        if (!audioRepository.existsById(id)) {
+            new DocumentationResponse("Item Not Found.");
+        }
+
+        //return audioRepository.findById(id).orElseThrow(() -> new DocumentationResponse("Item Not Found"));
+
         return null;
     }
 
@@ -88,9 +111,5 @@ public class AudioServiceImpl {
 
     public List<Audio> findByRecordLabel(String recordLabel) {
         return null;
-    }
-
-    public List<Audio> listAllAudioItems() {
-        return audioRepository.findAll();
     }
 }
